refactored: ServerRunner(SikulixServer): renamed class name, changed package, organized import, removed extends AbstractScriptRunner
