+++ API/src/main/java/org/sikuli/script/support/RunTime.java
@@ -11,7 +11,6 @@ import org.sikuli.natives.WinUtil;
 import org.sikuli.script.*;
 import org.sikuli.script.runnerHelpers.JythonHelper;
 import org.sikuli.script.runners.JythonRunner;
-import org.sikuli.script.runners.ServerRunner;
 import org.sikuli.util.CommandArgs;
 import org.sikuli.util.CommandArgsEnum;
 import org.sikuli.script.runners.ProcessRunner;
@@ -191,7 +190,7 @@ public class RunTime {
     }
 
     if (shouldRunServer()) {
-      if (ServerRunner.run(null)) {
+      if (SikulixServer.run(null)) {
         Sikulix.terminate(1, "");
       }
       Sikulix.terminate();
