+++ API/src/main/resources/htdocs/ControlBox.html
@@ -36,7 +36,7 @@
       </section>
     </main>
     <footer>
-      <a href="https://sikulix-2014.readthedocs.io/en/latest/scenarios.html#experimental-runserver-run-scripts-from-anywhere-with-zero-delay" target="_blank">SikuliX:ServerRunner Guide</a>
+      <a href="https://sikulix-2014.readthedocs.io/en/latest/scenarios.html#experimental-runserver-run-scripts-from-anywhere-with-zero-delay" target="_blank">SikulixServer Guide</a>
       <div>
         <p>&copy; 2019 sikuli.org, sikulix.com - MIT license</p>
       </div>
